Add files via upload
The creation of a comprehensive face recognition system that can operate on any kind of picture and is always becoming better was our aim. This enhancement has to be self-sufficient in order for it to detect the persons it already contained more accurately and to add new ones. Additionally, because this recognition must be performed as near to in real-time as feasible, the time requirements were a problem. The goal of this research was to create a model that could detect and evaluate toddler behaviors linked to ASD. Many of the challenges that children experience, including those related to communication, social skills, and learning problems, may be alleviated with early diagnosis of ASD. Using convolutional neural networks (CNNs) and performance assessment measures, the research examined children's non-clinical datasets.